move UniqueInstanceMixin to utils
